requirements.txt
smolagents[litellm]
playwright
Pillow
python-dotenv
click
openai
